Restructured java packages
