+++ src/main/java/com/rizwan/quotesapp/config/MongoConfig.java
@@ -4,6 +4,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 @Configuration
-@EnableMongoRepositories(basePackages = "com.rizwan.quotesapp.repository")
+@EnableMongoRepositories(basePackages = "com.rizwan.quotesapp.quote.repository")
 public class MongoConfig {
 }
