--- src/main/java/com/rizwan/quotesapp/model/enumeration/CreationType.java
@@ -1,6 +1,0 @@
-package com.rizwan.quotesapp.model.enumeration;
-
-public enum CreationType {
-    MANUAL,
-    SAVED
-}
