+++ src/main/java/com/rizwan/quotesapp/quote/model/enumeration/CreationType.java
@@ -1,0 +1,6 @@
+package com.rizwan.quotesapp.quote.model.enumeration;
+
+public enum CreationType {
+    MANUAL,
+    SAVED
+}
