+++ src/main/java/com/rizwan/quotesapp/quote/repository/QuoteRepository.java
@@ -1,6 +1,6 @@
-package com.rizwan.quotesapp.repository;
+package com.rizwan.quotesapp.quote.repository;
 
-import com.rizwan.quotesapp.model.Quote;
+import com.rizwan.quotesapp.quote.model.Quote;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
